+++ exoline/exo.py
@@ -553,6 +553,30 @@ class ExoConfig:
                 self.config = {}
 
     def lookup_shortcut(self, cik):
+        '''If a CIK has client/resource parts, seperate and look thouse up'''
+        if ':c' in cik:
+            # break into parts, then lookup each.
+            c,g,r = cik.partition(':c')
+            cik = { 'cik': self._lookup_shortcut(c),
+                    'client_id': self._lookup_shortcut(r) }
+        elif ':r' in cik:
+            c,g,r = cik.partition(':r')
+            cik = { 'cik': self._lookup_shortcut(c),
+                    'resource_id': self._lookup_shortcut(r) }
+        else:
+            # look it up, then check again for parts.
+            cik = self._lookup_shortcut(cik)
+            if ':c' in cik:
+                c,g,r = cik.partition(':c')
+                cik = {'cik': c, 'client_id': r}
+            elif ':r' in cik:
+                c,g,r = cik.partition(':r')
+                cik = {'cik': c, 'resource_id': r}
+
+        return cik
+
+
+    def _lookup_shortcut(self, cik):
         '''Look up what was passed for cik in config file
             if it doesn't look like a CIK.'''
         if self.regex_rid.match(cik) is None:
@@ -603,6 +627,16 @@ class ExolineOnepV1(onep.OnepV1):
             te = time.time()
             PERF_DATA.append({'cik': cik, 'procedures': procedures, 'seconds': te-ts})
         return r
+
+    def _getAuth(self, cik):
+        '''Definition of what is passed for cik is different that what OnepV1 expects.
+        so adjust to match.'''
+        if type(cik) is dict:
+            return cik
+        else:
+            return {"cik": cik}
+
+        return cik
 
     def comment(self, cik, rid, visibility, comment, defer=False):
         return self._call('comment', cik, [rid, visibility, comment], defer)